moving to broken out unit tests
